[map-helper-functions] addUIDs helper function
