+++ packages/map/src/helpers/index.ts
@@ -18,3 +18,4 @@ export { sortSpecialClassesLast } from './sortSpecialClassesLast'
 export { titleCase } from './titleCase'
 export { titleCase as toTitleCase } from './toTitleCase'
 export { validateFipsCodeLength } from './validateFipsCodeLength'
+export { addUIDs } from './addUIDs'
